% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brooklyn99_palette.R
\name{brooklyn99_dark_pal}
\alias{brooklyn99_dark_pal}
\title{brooklyn99_dark_pal}
\usage{
brooklyn99_dark_pal()
}
\value{
B99 dark palette
}
\description{
dark version of Brooklyn Nine-Nine palette
}
\seealso{
\code{\link[scales]{manual_pal}}
}
