% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brooklyn99_palette.R
\name{brooklyn99_pal}
\alias{brooklyn99_pal}
\title{brooklyn99_pal}
\usage{
brooklyn99_pal()
}
\description{
Brooklyn Nine Nine Color and Fill Scales
}
\details{
Colors that work well with the blue background!
}
\seealso{
\code{\link[scales]{manual_pal}}
}
