% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_champion_htf_heavyweight}
\alias{import_champion_htf_heavyweight}
\title{import_champion_htf_heavyweight}
\usage{
import_champion_htf_heavyweight()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
imports Champion HTF-Heavyweight font
}
\details{
import_*() functions inspired by hrbrmstrthemes
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".
}
\seealso{
\code{\link[extrafont]{font_import}}
}
