% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_ChelseaMarket}
\alias{import_ChelseaMarket}
\title{import_chelseaMarket}
\usage{
import_ChelseaMarket()
}
\description{
Imports Chelsea Market (Hilda)
}
\details{
import_*() functions taken from hrbrthemes.
}
\seealso{
\code{\link[extrafont]{font_import}}
}
