% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_cinzel}
\alias{import_cinzel}
\title{import_cinzel}
\usage{
import_cinzel()
}
\description{
Game of Thrones font ("Cinzel" font)
}
\details{
import_*() functions inspired by hrbrmstrthemes, actual font is "Trajan Pro".
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".
}
\seealso{
\code{\link[extrafont]{font_import}}
}
