% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_rickAndMorty}
\alias{import_rickAndMorty}
\title{import_rickAndMorty}
\usage{
import_rickAndMorty()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Rick & Morty font ("Get Schwifty")
}
\details{
Actual font is ... well, Justin Roiland's actual handwriting
}
