% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_roboto_condensed}
\alias{import_roboto_condensed}
\title{import_roboto_condensed}
\usage{
import_roboto_condensed()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
taken from hrbrmstrthemes
}
\details{
DETAILS
}
\seealso{
\code{\link[extrafont]{font_import}}
}
