% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_robotoCondensed}
\alias{import_robotoCondensed}
\title{import_robotoCondensed}
\usage{
import_robotoCondensed()
}
\description{
taken from hrbrthemes
}
\details{
import_*() functions taken from hrbrthemes.
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".
}
\seealso{
\code{\link[extrafont]{font_import}}
}
