% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_simpsons}
\alias{import_simpsons}
\title{import_simpsons}
\usage{
import_simpsons()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The Simpsons Font ("Akbar" font)
}
\details{
DETAILS
}
\seealso{
\code{\link[extrafont]{font_import}}
}
