% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_spongeBob}
\alias{import_spongeBob}
\title{import_spongeBob}
\usage{
import_spongeBob()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
spongeBob SquarePants font ("Some-Time-Later")
}
\details{
DETAILS
}
