% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_titillium_web}
\alias{import_titillium_web}
\title{import_titillium_web}
\usage{
import_titillium_web()
}
\description{
Imports Titillium Web
}
\details{
import_*() functions taken from hrbrthemes.
}
\seealso{
\code{\link[extrafont]{font_import}}
}
