% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_titilliumWeb}
\alias{import_titilliumWeb}
\title{Import "Titillium Web" font}
\usage{
import_titilliumWeb()
}
\description{
Imports Titillium Web
}
\details{
import_*() functions taken from hrbrthemes.
}
\seealso{
\code{\link[extrafont]{font_import}}
}
