% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_univers}
\alias{import_univers}
\title{import_univers}
\usage{
import_univers()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
import 'Univers' font
}
\details{
import_*() functions inspired by hrbrmstrthemes
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".
}
\examples{
\dontrun{
if(interactive()){
 import_univers()
 }
}
}
\seealso{
\code{\link[extrafont]{font_import}}
}
