% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_univers_extrablack}
\alias{import_univers_extrablack}
\title{import_univers_extrablack}
\usage{
import_univers_extrablack()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[extrafont]{font_import}}
}
