% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bikinibottom.R
\name{paint_bikiniBottom}
\alias{paint_bikiniBottom}
\title{paint_bikiniBottom}
\usage{
paint_bikiniBottom(sponge_gg, width = 800, height = 500,
  outfile = NULL, sponge_background = sponge_images(c("background",
  "floral", "floral2")), ...)
}
\arguments{
\item{sponge_gg}{ggplot object}

\item{width}{PARAM_DESCRIPTION, Default: 800}

\item{height}{PARAM_DESCRIPTION, Default: 500}

\item{outfile}{PARAM_DESCRIPTION, Default: NULL}

\item{sponge_background}{PARAM_DESCRIPTION, Default: sponge_images(c("background", "floral", "floral2"))}

\item{...}{PARAM_DESCRIPTION}

\item{pointsize}{PARAM_DESCRIPTION, Default: 16}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
add SpongeBob background
}
\details{
Adapted from ggpomological's paint_pomological() function!
}
\seealso{
\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
 \code{\link[magick]{device}},\code{\link[magick]{editing}},\code{\link[magick]{transform}},\code{\link[magick]{composite}}
}
