% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikinibottom.R
\name{paintBikiniBottom}
\alias{paintBikiniBottom}
\title{paintBikiniBottom}
\usage{
paintBikiniBottom(sponge.gg, width = 800, height = 500,
  output.file = NULL, sponge.background = sponge_images(c("background",
  "floral")), ...)
}
\arguments{
\item{sponge.gg}{ggplot object}

\item{width}{PARAM_DESCRIPTION, Default: 800}

\item{height}{PARAM_DESCRIPTION, Default: 500}

\item{output.file}{PARAM_DESCRIPTION, Default: NULL}

\item{sponge.background}{PARAM_DESCRIPTION, Default: sponge_images(c("background", "floral", "floral2"))}

\item{...}{PARAM_DESCRIPTION}
}
\value{
Your plot with a Spongebob themed background!
}
\description{
add SpongeBob background
}
\details{
Adapted from ggpomological's paint_pomological() function!
}
