% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_color_airNomads}
\alias{scale_color_airNomads}
\title{scale_color_airNomads}
\usage{
scale_color_airNomads(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Air Nomads color scale
}
\details{
DETAILS
}
