% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brooklyn99_palette.R
\name{scale_color_brooklyn99_dark}
\alias{scale_color_brooklyn99_dark}
\title{scale_color_brooklyn99_dark}
\usage{
scale_color_brooklyn99_dark(...)
}
\arguments{
\item{...}{any other argument that can go into scale_color_*()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
DARK color scale B99
}
\details{
DETAILS
}
