% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_color_earthKingdom}
\alias{scale_color_earthKingdom}
\title{scale_color_earthKingdom}
\usage{
scale_color_earthKingdom(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
Earth Kingdom color scale
}
