% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_color_fireNation}
\alias{scale_color_fireNation}
\title{scale_color_fireNation}
\usage{
scale_color_fireNation(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
Fire Nation color scale
}
