% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_color_greyjoy}
\alias{scale_color_greyjoy}
\title{scale_color_greyjoy}
\usage{
scale_color_greyjoy(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
House Greyjoy color scale
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
