% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_color_lannister}
\alias{scale_color_lannister}
\title{scale_color_lannister}
\usage{
scale_color_lannister(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
House Lannister color scale
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
