% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parksAndRec_palette.R
\name{scale_color_parksAndRec}
\alias{scale_color_parksAndRec}
\title{scale_color_parksAndRec}
\usage{
scale_color_parksAndRec(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
Parks & Recreation color scale
}
