% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickandmorty_palette.R
\name{scale_color_rickAndMorty}
\alias{scale_color_rickAndMorty}
\title{scale_color_rickAndMorty}
\usage{
scale_color_rickAndMorty(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Rick & Morty color scale
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
