% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsons_palette.R
\name{scale_color_simpsons}
\alias{scale_color_simpsons}
\title{scale_color_simpsons}
\usage{
scale_color_simpsons(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
The Simpsons color scale
}
