% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongebob_palette.R
\name{scale_color_spongeBob}
\alias{scale_color_spongeBob}
\title{scale_color_spongeBob}
\usage{
scale_color_spongeBob(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
Spongebob Squarepants color scale
}
