% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_color_stark}
\alias{scale_color_stark}
\title{scale_color_stark}
\usage{
scale_color_stark(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
House Stark color scale
}
\details{
DETAILS
}
