% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_color_targaryen}
\alias{scale_color_targaryen}
\title{scale_color_targaryen}
\usage{
scale_color_targaryen(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
House Targaryen color scale
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
