% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_color_tully}
\alias{scale_color_tully}
\title{scale_color_tully}
\usage{
scale_color_tully(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
House Tully color scale
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
