% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_color_waterTribe}
\alias{scale_color_waterTribe}
\title{scale_color_waterTribe}
\usage{
scale_color_waterTribe(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Water Tribe color scale
}
\details{
DETAILS
}
