% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brooklyn99_palette.R
\name{scale_colour_brooklyn99}
\alias{scale_colour_brooklyn99}
\title{scale_colour_brooklyn99}
\usage{
scale_colour_brooklyn99(...)
}
\arguments{
\item{...}{any other argument that can go into scale_colour_*()}
}
\description{
for British-spelling compatibility
}
