% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brooklyn99_palette.R
\name{scale_colour_brooklyn99_dark}
\alias{scale_colour_brooklyn99_dark}
\title{scale_colour_brooklyn99_dark}
\usage{
scale_colour_brooklyn99_dark(...)
}
\description{
for British-spelling compatibility
}
