% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_colour_fireNation}
\alias{scale_colour_fireNation}
\title{scale_colour_fireNation}
\usage{
scale_colour_fireNation(...)
}
\arguments{
\item{...}{any other argument that can go into scale_colour_*()}
}
\description{
for British-spelling compatibility
}
