% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_colour_greyjoy}
\alias{scale_colour_greyjoy}
\title{scale_colour_grey}
\usage{
scale_colour_greyjoy(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
House Greyjoy colour scale
}
