% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_colour_lannister}
\alias{scale_colour_lannister}
\title{scale_colour_lannister}
\usage{
scale_colour_lannister(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
