% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickandmorty_palette.R
\name{scale_colour_rickAndMorty}
\alias{scale_colour_rickAndMorty}
\title{scale_colour_rickAndMorty}
\usage{
scale_colour_rickAndMorty(...)
}
\arguments{
\item{...}{any other argument that can go into scale_colour_*()}
}
\description{
for British-spelling compatibility
}
