% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsons_palette.R
\name{scale_colour_simpsons}
\alias{scale_colour_simpsons}
\title{scale_colour_simpsons}
\usage{
scale_colour_simpsons(...)
}
\arguments{
\item{...}{any other argument that can go into scale_colour_*()}
}
\description{
for British-spelling compatibility
}
