% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongebob_palette.R
\name{scale_colour_spongeBob}
\alias{scale_colour_spongeBob}
\title{scale_colour_spongeBob}
\usage{
scale_colour_spongeBob(...)
}
\arguments{
\item{...}{any other argument that can go into scale_colour_*()}
}
\description{
for British-spelling compatibility
}
