% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_colour_targaryen}
\alias{scale_colour_targaryen}
\title{scale_colour_targaryen}
\usage{
scale_colour_targaryen(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
House Targaryen colour scale
}
