% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_colour_tully}
\alias{scale_colour_tully}
\title{scale_colour_tully}
\usage{
scale_colour_tully(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
