% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_colour_tyrell}
\alias{scale_colour_tyrell}
\title{scale_colour_tyrell}
\usage{
scale_colour_tyrell(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
FUNCTION_DESCRIPTION
}
