% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_fill_air_nomads}
\alias{scale_fill_air_nomads}
\title{scale_fill_air_nomads}
\usage{
scale_fill_air_nomads(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Air Nomads fill scale
}
\details{
DETAILS
}
