% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_fill_airNomads}
\alias{scale_fill_airNomads}
\title{scale_fill_air_nomads}
\usage{
scale_fill_airNomads(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
Air Nomads fill scale
}
