% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brooklyn99_palette.R
\name{scale_fill_brooklyn99}
\alias{scale_fill_brooklyn99}
\title{scale_fill_brooklyn99}
\usage{
scale_fill_brooklyn99(...)
}
\arguments{
\item{...}{misc. arguments that can go into scale_fill_*() functions}
}
\description{
fill scales for Brooklyn Nine-Nine
}
