% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brooklyn99_palette.R
\name{scale_fill_brooklyn99_dark}
\alias{scale_fill_brooklyn99_dark}
\title{scale_fill_brooklyn99_dark}
\usage{
scale_fill_brooklyn99_dark(...)
}
\arguments{
\item{...}{misc. arguments that can go into scale_fill_*() functions}
}
\description{
DARK fill scales for Brooklyn Nine-Nine
}
