% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_fill_earth_kingdom}
\alias{scale_fill_earth_kingdom}
\title{scale_fill_earth_kingdom}
\usage{
scale_fill_earth_kingdom(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Earth Kingdom fill scale
}
\details{
DETAILS
}
