% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_fill_fire_nation}
\alias{scale_fill_fire_nation}
\title{scale_fill_fire_nation}
\usage{
scale_fill_fire_nation(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Fire Nation fill scale
}
\details{
DETAILS
}
