% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parksAndRec_palette.R
\name{scale_fill_parksAndRec}
\alias{scale_fill_parksAndRec}
\title{scale_fill_parksAndRec}
\usage{
scale_fill_parksAndRec(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
Parks & Recreation fill scale
}
