% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickandmorty_palette.R
\name{scale_fill_rickAndMorty}
\alias{scale_fill_rickAndMorty}
\title{scale_fill_rickAndMorty}
\usage{
scale_fill_rickAndMorty(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
Rick & Morty fill scale
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
