% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongebob_palette.R
\name{scale_fill_spongeBob}
\alias{scale_fill_spongeBob}
\title{scale_fill_spongeBob}
\usage{
scale_fill_spongeBob(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
Spongebob Squarepants fill scale
}
