% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones_palette.R
\name{scale_fill_stark}
\alias{scale_fill_stark}
\title{scale_fill_stark}
\usage{
scale_fill_stark(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
House Stark fill scale
}
\details{
DETAILS
}
