% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_fill_water_tribe}
\alias{scale_fill_water_tribe}
\title{scale_fill_water_tribe}
\usage{
scale_fill_water_tribe(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Water Tribe fill scale
}
\details{
DETAILS
}
