% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelastairbender_palette.R
\name{scale_fill_waterTribe}
\alias{scale_fill_waterTribe}
\title{scale_fill_waterTribe}
\usage{
scale_fill_waterTribe(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\description{
Water Tribe fill scale
}
