% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bikinibottom.R
\name{sponge_images}
\alias{sponge_images}
\title{sponge_images}
\usage{
sponge_images(which = c("background", "floral", "floral2"))
}
\arguments{
\item{which}{PARAM_DESCRIPTION, Default: c("background", "floral", "floral2")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
find SpongeBob background images
}
\details{
DETAILS
}
