% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hilda.R
\name{theme_hildaDusk}
\alias{theme_hildaDusk}
\title{Hilda "Dusk" theme}
\usage{
theme_hildaDusk(
  text.font = "Chelsea Market",
  title.font = "Chelsea Market",
  legend.font = "Chelsea Market",
  title.size = 18,
  text.size = 14,
  subtitle.size = 12,
  axis.title.size = 14,
  axis.text.size = 12,
  legend.title.size = 10,
  legend.text.size = 9,
  title.color = "#F9FEFF",
  subtitle.color = "#F9FEFF",
  text.color = "#F9FEFF",
  axis.title.color = "#F9FEFF",
  axis.text.color = "#F9FEFF",
  legend.title.color = "#F9FEFF",
  legend.text.color = "#F9FEFF",
  legend.position = "bottom",
  ticks = FALSE
)
}
\arguments{
\item{text.font}{text font, Default: "Chelsea Market"}

\item{title.font}{title font, Default: "Chelsea Market"}

\item{legend.font}{legend font, Default: "Chelsea Market"}

\item{title.size}{title font size, Default: 18}

\item{text.size}{text font size, Default: 14}

\item{subtitle.size}{subtitle font size, Default: 12}

\item{axis.title.size}{axis title font size, Default: 14}

\item{axis.text.size}{axis text font size, Default: 12}

\item{legend.title.size}{legend title font size, Default: 10}

\item{legend.text.size}{legend text font size, Default: 9}

\item{title.color}{title color, Default: '#F9FEFF'}

\item{subtitle.color}{subtitle color, Default: '#F9FEFF'}

\item{text.color}{text color, Default: '#F9FEFF'}

\item{axis.title.color}{axis title color, Default: '#F9FEFF'}

\item{axis.text.color}{axis text color, Default: '#F9FEFF'}

\item{legend.title.color}{legend title color, Default: '#F9FEFF'}

\item{legend.text.color}{legend text color, Default: '#F9FEFF'}

\item{legend.position}{legend position, Default: 'bottom'}

\item{ticks}{add axis ticks, Default: FALSE}
}
\description{
Hilda theme
}
\examples{
library(ggplot2)

ggplot(airquality, aes(x = Day, y = Temp,
     group = as.factor(Month), color = as.factor(Month))) +
     geom_point(size = 2.5) +
     scale_color_hilda(palette = "Dusk") +
     theme_hildaDusk(text.font = "Times", title.font = "Times",
                     legend.font = "Times")
}
