% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_parksAndRec.R
\name{theme_parksAndRec}
\alias{theme_parksAndRec}
\title{theme_parksAndRec}
\usage{
theme_parksAndRec(font.text = "Champion HTF-Heavyweight",
  font.title = "Champion HTF-Heavyweight",
  font.legend = "Champion HTF-Heavyweight")
}
\arguments{
\item{font.text}{text font}

\item{font.title}{title font}

\item{font.legend}{legend font}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Parks & Recreation theme
}
\details{
DETAILS
}
