% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rickandmorty.R
\name{theme_rickAndMorty}
\alias{theme_rickAndMorty}
\title{theme_rickAndMorty}
\usage{
theme_rickAndMorty(font.text = NULL, font.title = NULL,
  font.legend = NULL, title.size = 20, text.size = 12,
  subtitle.size = 14, axis.title.size = 14, axis.text.size = 10,
  legend.title.size = 10, legend.text.size = 9, title.color = NULL,
  subtitle.color = NULL, text.color = NULL, axis.title.color = NULL,
  axis.text.color = NULL, legend.title.color = NULL,
  legend.text.color = NULL)
}
\arguments{
\item{font.text}{text font}

\item{font.title}{title font}

\item{font.legend}{legend font}

\item{title.size}{title size}

\item{text.size}{text font size}

\item{subtitle.size}{subtitle font size}

\item{axis.title.size}{axis title font size}

\item{axis.text.size}{axis text font size}

\item{legend.title.size}{legend title font size}

\item{legend.text.size}{legend text font size}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Rick & Morty theme, Font: "Get Schwifty"
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
