% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_thelastairbender.R
\name{theme_theLastAirbender}
\alias{theme_theLastAirbender}
\title{theme_theLastAirbender}
\usage{
theme_theLastAirbender(font.text = "Slayer", font.title = "Slayer",
  font.legend = "Slayer", title.size = 14, text.size = 10,
  subtitle.size = 12, axis.title.size = 10, axis.text.size = 8,
  legend.title.size = 10, legend.text.size = 8)
}
\arguments{
\item{font.text}{PARAM_DESCRIPTION, Default: 'Slayer'}

\item{font.title}{PARAM_DESCRIPTION, Default: 'Slayer'}

\item{font.legend}{PARAM_DESCRIPTION, Default: 'Slayer'}

\item{title.size}{title font size}

\item{text.size}{text font size}

\item{subtitle.size}{subtitle font size}

\item{axis.title.size}{axis title font size}

\item{axis.text.size}{axis text font size}

\item{legend.title.size}{legend title font size}

\item{legend.text.size}{legend text font size}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Avatar: The Last Airbender theme
}
\details{
DETAILS
}
