% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvthemes-package.R
\docType{package}
\name{tvthemes}
\alias{tvthemes}
\alias{tvthemes-package}
\title{TV Show Themes and Color Palettes for 'ggplot2' Graphics}
\description{
Contains various 'ggplot2' themes and color palettes based on TV shows.
}
\details{
\itemize{
\item Avatar: The Last Airbender: theme + palettes (Fire Nation, Water Tribe,
Earth Kingdom, & Air Nomads)
\item Brooklyn Nine-Nine: theme + palettes (regular & dark)
\item Game of Thrones/A Song of Ice & Fire: ‘The Palettes of Ice & Fire’
(currently: Stark, Baratheon (Stannis), Lannister, Tully, Targaryen, Martell,
Greyjoy, Tyrell, Arryn, & Manderly)
\item Rick & Morty: theme + palette
\item Parks & Recreation: two themes (light & dark) + palette
\item The Simpsons: theme + palette
\item Spongebob Squarepants: theme + palette + background images
\item Hilda (Day, Dusk, Night) themes + palettes
\item Attack on Titan: palette
\item Kim Possible: palette
\item Big Hero 6: palette
}

Most palettes are geared toward 'discrete' scales rather than 'continuous' so
the option isn't available in the functions. This may change in the future.
}
\seealso{
\itemize{
\item https://ryo-n7.github.io/2019-05-16-introducing-tvthemes-package/
\item https://github.com/Ryo-N7/tvthemes
}
}
\author{
Ryo Nakagawara (@R_by_Ryo)
}
\keyword{internal}
