% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_champion_htf_heaveyweight}
\alias{import_champion_htf_heaveyweight}
\title{import_champion_htf_heaveyweight}
\usage{
import_champion_htf_heaveyweight()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
imports Champion HTF-Heavyweight font
}
\details{
DETAILS
}
\seealso{
\code{\link[extrafont]{font_import}}
}
