% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_chelseaMarket}
\alias{import_chelseaMarket}
\title{import_chelseaMarket}
\usage{
import_chelseaMarket()
}
\description{
Imports Chelsea Market (Hilda)
}
\details{
import_*() functions taken from hrbrthemes.
Formerly `import_ChelseaMarket()`.
}
\seealso{
\code{\link[extrafont]{font_import}}
}
