% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_roboto_condensed}
\alias{import_roboto_condensed}
\title{import_roboto_condensed}
\usage{
import_roboto_condensed()
}
\description{
taken from hrbrmstrthemes
}
\details{
import_*() functions inspired by hrbrmstrthemes
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".
}
\seealso{
\code{\link[extrafont]{font_import}}
}
