% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_roboto_condensed}
\alias{import_roboto_condensed}
\title{Import "Roboto Condensed" font (deprecated)}
\usage{
import_roboto_condensed()
}
\description{
taken from hrbrthemes
}
\details{
import_*() functions taken from hrbrthemes.
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".
}
\seealso{
\code{\link[extrafont]{font_import}}
}
