% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_simpsons}
\alias{import_simpsons}
\title{import_simpsons}
\usage{
import_simpsons()
}
\description{
The Simpsons Font ("Akbar" font)
}
\details{
import_*() functions inspired by hrbrmstrthemes
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".
}
\seealso{
\code{\link[extrafont]{font_import}}
}
