% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_southpark}
\alias{import_southpark}
\title{import_southpark}
\usage{
import_southpark()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
South Park font
}
\details{
import_*() functions inspired by hrbrmstrthemes
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[extrafont]{font_import}}
}
