% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_theLastAirbender}
\alias{import_theLastAirbender}
\title{import_theLastAirbender}
\usage{
import_theLastAirbender()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The Last Airbender font ("Slayer")
}
\details{
Actual font is Herculanum (can be bought for $35)
}
