% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.r
\name{import_theLastAirbender}
\alias{import_theLastAirbender}
\title{Import "Slayer" font (deprecated)}
\usage{
import_theLastAirbender()
}
\description{
The Last Airbender font ("Slayer")
}
\details{
Actual font is Herculanum.
import_*() functions taken from hrbrthemes.
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".
}
