% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikinibottom.R
\name{paintBikiniBottom}
\alias{paintBikiniBottom}
\title{Add SpongeBob background}
\usage{
paintBikiniBottom(
  plot,
  width = 800,
  height = 500,
  output.file = NULL,
  background = "background",
  ...
)
}
\arguments{
\item{plot}{the ggplot object you want to Spongbobify!}

\item{width}{width, Default: 800}

\item{height}{height, Default: 500}

\item{output.file}{File path to save image, Default: NULL}

\item{background}{"background" or "floral", Default: "background"}

\item{...}{Other options, see `?magick::image_graph()`}
}
\value{
Your plot with a Spongebob themed background!
}
\description{
Add SpongeBob background
}
\details{
Adapted from ggpomological's `paint_pomological()` function!
}
