% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikinibottom.R
\name{sponge_images}
\alias{sponge_images}
\title{sponge_images}
\usage{
sponge_images(which = c("background", "floral"))
}
\arguments{
\item{which}{Choose image, Default: c("background", "floral")}
}
\description{
find SpongeBob background images
}
\keyword{internal}
