% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_brooklyn99.R
\name{theme_brooklyn99}
\alias{theme_brooklyn99}
\title{theme_brooklyn99}
\usage{
theme_brooklyn99(text.font = NULL, title.font = NULL,
  legend.font = NULL, title.size = 18, text.size = 14,
  subtitle.size = 12, axis.title.size = 14, axis.text.size = 12,
  legend.title.size = 10, legend.text.size = 9,
  title.color = "#F9FEFF", subtitle.color = "#F9FEFF",
  text.color = "#F9FEFF", axis.title.color = "#F9FEFF",
  axis.text.color = "#F9FEFF", legend.title.color = "#F9FEFF",
  legend.text.color = "#F9FEFF", legend.position = "bottom")
}
\arguments{
\item{text.font}{text font}

\item{title.font}{title font}

\item{legend.font}{legend font}

\item{title.size}{title font size, Default: 18}

\item{text.size}{text font size, Default: 14}

\item{subtitle.size}{subtitle font size, Default: 12}

\item{axis.title.size}{axis title font size, Default: 14}

\item{axis.text.size}{axis text font size, Default: 12}

\item{legend.title.size}{legend title font size, Default: 10}

\item{legend.text.size}{legend text font size, Default: 9}

\item{title.color}{title color, Default: "F9FEFF"}

\item{subtitle.color}{subtitle.color, Default: "F9FEFF"}

\item{text.color}{text color, Default: "F9FEFF"}

\item{axis.title.color}{axis title color, Default: "F9FEFF"}

\item{axis.text.color}{axis text color, Default: "F9FEFF"}

\item{legend.title.color}{legend title color, Default: "F9FEFF"}

\item{legend.text.color}{legend text color, Default: "F9FEFF"}

\item{legend.position}{legend position, Default: "bottom"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Brooklyn Nine-Nine theme, Recommended font: "Roboto Condensed" (title), "Calibri Light" (other text)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 ggplot(mtcars, aes(x = hp, y = mpg)) +
 geom_point() +
 theme_brooklyn99()
 }
}
}
\seealso{
[ggplot2::theme]
}
