% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_brooklyn99.R
\name{theme_brooklyn99}
\alias{theme_brooklyn99}
\title{theme_brooklyn99}
\usage{
theme_brooklyn99(font.text = "Calibri Light",
  font.title = "Calibri Bold", font.legend = "Calibri Light",
  title.size = 20)
}
\arguments{
\item{font.text}{main font text, Default: Univers}

\item{font.title}{title font, Default: Univers LT 93 ExtraBlackEx}

\item{font.legend}{legend font, Default: Univers}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Brooklyn Nine-Nine theme
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 ggplot(mtcars, aes(x = hp, y = mpg)) +
 geom_point() +
 theme_brooklyn99()
 }
}
}
\seealso{
[ggplot2::theme]
}
