% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_parksAndRec.R
\name{theme_parksAndRec}
\alias{theme_parksAndRec}
\title{theme_parksAndRec}
\usage{
theme_parksAndRec(font.text = NULL, font.title = NULL,
  font.legend = NULL, title.size = 20, text.size = 16,
  subtitle.size = 14, axis.title.size = 14, axis.text.size = 12,
  legend.title.size = 14, legend.text.size = 12, title.color = NULL,
  subtitle.color = NULL, text.color = NULL, axis.title.color = NULL,
  axis.text.color = NULL, legend.title.color = NULL,
  legend.text.color = NULL)
}
\arguments{
\item{font.text}{text font}

\item{font.title}{title font}

\item{font.legend}{legend font}

\item{title.size}{title font size}

\item{text.size}{text font size}

\item{subtitle.size}{subtitle font size}

\item{axis.title.size}{axis title font size}

\item{axis.text.size}{axis text font size}

\item{legend.title.size}{legend title font size}

\item{legend.text.size}{legend text font size}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Parks & Recreation theme, Recommended font: "Titillium Web"
}
\details{
DETAILS
}
