% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_parksAndRec.R
\name{theme_parksAndRec_light}
\alias{theme_parksAndRec_light}
\title{theme_parksAndRec_light}
\usage{
theme_parksAndRec_light(text.font = NULL, title.font = NULL,
  legend.font = NULL, title.size = 20, text.size = 16,
  subtitle.size = 14, axis.title.size = 14, axis.text.size = 12,
  legend.title.size = 14, legend.text.size = 12,
  title.color = "grey20", subtitle.color = "grey20",
  text.color = "grey20", axis.title.color = "grey20",
  axis.text.color = "grey20", legend.title.color = "grey20",
  legend.text.color = "grey20")
}
\arguments{
\item{text.font}{text font}

\item{title.font}{title font}

\item{legend.font}{legend font}

\item{title.size}{title font size}

\item{text.size}{text font size}

\item{subtitle.size}{subtitle font size}

\item{axis.title.size}{axis title font size}

\item{axis.text.size}{axis text font size}

\item{legend.title.size}{legend title font size}

\item{legend.text.size}{legend text font size}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Parks & Recreation light theme, Recommended font: "Titillium Web"
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
