% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_parksAndRec.R
\name{theme_parksAndRec_light}
\alias{theme_parksAndRec_light}
\title{theme_parksAndRec_light}
\usage{
theme_parksAndRec_light(font.text = "Titillium Web",
  font.title = "Titillium Web Black", font.legend = "Titillium Web",
  title.size = 20, text.size = 16, subtitle.size = 14,
  axis.title.size = 14, axis.text.size = 12, legend.title.size = 14,
  legend.text.size = 12)
}
\arguments{
\item{font.text}{PARAM_DESCRIPTION, Default: 'Champion HTF-Heavyweight'}

\item{font.title}{PARAM_DESCRIPTION, Default: 'Champion HTF-Heavyweight'}

\item{font.legend}{PARAM_DESCRIPTION, Default: 'Champion HTF-Heavyweight'}

\item{title.size}{title font size}

\item{text.size}{text font size}

\item{subtitle.size}{subtitle font size}

\item{axis.title.size}{axis title font size}

\item{axis.text.size}{axis text font size}

\item{legend.title.size}{legend title font size}

\item{legend.text.size}{legend text font size}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Parks & Recreation light theme
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
