% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_parksAndRec.R
\name{theme_parksAndRec_light}
\alias{theme_parksAndRec_light}
\title{theme_parksAndRec_light}
\usage{
theme_parksAndRec_light(font.text = "Champion HTF-Heavyweight",
  font.title = "Champion HTF-Heavyweight",
  font.legend = "Champion HTF-Heavyweight")
}
\arguments{
\item{font.text}{PARAM_DESCRIPTION, Default: 'Champion HTF-Heavyweight'}

\item{font.title}{PARAM_DESCRIPTION, Default: 'Champion HTF-Heavyweight'}

\item{font.legend}{PARAM_DESCRIPTION, Default: 'Champion HTF-Heavyweight'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Parks & Recreation light theme
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
