% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rickandmorty.R
\name{theme_rickAndMorty}
\alias{theme_rickAndMorty}
\title{theme_rickAndMorty}
\usage{
theme_rickAndMorty(font.text = "Get Schwifty",
  font.title = "Get Schwifty", font.legend = "Get Schwifty",
  title.size = 18)
}
\arguments{
\item{font.text}{text font, Default: 'Get Schwifty'}

\item{font.title}{title font, Default: 'Get Schwifty'}

\item{font.legend}{legend font, Default: 'Get Schwifty'}

\item{title.size}{title size, Default: 11}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Rick & Morty theme
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
