% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simpsons.R
\name{theme_simpsons}
\alias{theme_simpsons}
\title{theme_simpsons}
\usage{
theme_simpsons(text.font = NULL, title.font = NULL,
  legend.font = NULL, title.size = 18, text.size = 14,
  subtitle.size = 12, axis.title.size = 14, axis.text.size = 10,
  legend.title.size = 10, legend.text.size = 9,
  title.color = "#FFD235", subtitle.color = "#fee8c8",
  text.color = "#fee8c8", axis.title.color = "#fee8c8",
  axis.text.color = "#fee8c8", legend.title.color = "#ffffff",
  legend.text.color = "#ffffff")
}
\arguments{
\item{text.font}{text font}

\item{title.font}{title font}

\item{legend.font}{legend font}

\item{title.size}{title font size}

\item{text.size}{text font size}

\item{subtitle.size}{subtitle font size}

\item{axis.title.size}{axis title font size}

\item{axis.text.size}{axis text font size}

\item{legend.title.size}{legend title font size}

\item{legend.text.size}{legend text font size}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The Simpsons theme, Recommended font: "Akbar"
}
\details{
Largely inspired by `@nathancunn`'s blog posts on The Simpsons
}
