% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simpsons.R
\name{theme_simpsons}
\alias{theme_simpsons}
\title{theme_simpsons}
\usage{
theme_simpsons(font.text = "Akbar", font.title = "Akbar",
  font.legend = "Akbar", title.size = 11)
}
\arguments{
\item{font.text}{PARAM_DESCRIPTION, Default: 'Akbar'}

\item{font.title}{PARAM_DESCRIPTION, Default: 'Akbar'}

\item{font.legend}{PARAM_DESCRIPTION, Default: 'Akbar'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The Simpsons theme
}
\details{
DETAILS
}
