% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simpsons.R
\name{theme_simpsons}
\alias{theme_simpsons}
\title{theme_simpsons}
\usage{
theme_simpsons(font.text = "Akbar", font.title = "Akbar",
  font.legend = "Akbar", title.size = 18)
}
\arguments{
\item{font.text}{font for text, Default: 'Akbar'}

\item{font.title}{font for title, Default: 'Akbar'}

\item{font.legend}{font for legend, Default: 'Akbar'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The Simpsons theme
}
\details{
Largely inspired by `@nathancunn`'s blog posts
}
