% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_spongebob.R
\name{theme_spongeBob}
\alias{theme_spongeBob}
\title{theme_spongeBob}
\usage{
theme_spongeBob(font.text = "Some Time Later",
  font.title = "Some Time Later", font.legend = "Some Time Later",
  title.size = 14)
}
\arguments{
\item{font.text}{text font, Default: 'Some Time Later'}

\item{font.title}{title font, Default: 'Some Time Later'}

\item{font.legend}{legend font, Default: 'Some Time Later'}

\item{title.size}{size of title, Default: 14}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Spongebob Squarepants theme
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
