% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_spongebob.R
\name{theme_spongeBob}
\alias{theme_spongeBob}
\title{theme_spongeBob}
\usage{
theme_spongeBob(text.font = NULL, title.font = NULL,
  legend.font = NULL, title.size = 18, text.size = 12,
  subtitle.size = 12, axis.title.size = 14, axis.text.size = 12,
  legend.title.size = 10, legend.text.size = 9,
  title.color = "#F9FEFF", subtitle.color = "#F9FEFF",
  text.color = "#F9FEFF", axis.title.color = NULL,
  axis.text.color = "#F9FEFF", legend.title.color = "#F9FEFF",
  legend.text.color = "#F9FEFF")
}
\arguments{
\item{text.font}{text font}

\item{title.font}{title font}

\item{legend.font}{legend font}

\item{title.size}{size of title, Default: 14}

\item{text.size}{text font size}

\item{subtitle.size}{subtitle font size}

\item{axis.title.size}{axis title font size}

\item{axis.text.size}{axis text font size}

\item{legend.title.size}{legend title font size}

\item{legend.text.size}{legend text font size}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Spongebob Squarepants theme, Recommended font: "Some Time Later"
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
