% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_thelastairbender.R
\name{theme_theLastAirbender}
\alias{theme_theLastAirbender}
\title{theme_theLastAirbender}
\usage{
theme_theLastAirbender(text.font = NULL, title.font = NULL,
  legend.font = NULL, title.size = 14, text.size = 10,
  subtitle.size = 12, axis.title.size = 10, axis.text.size = 8,
  legend.title.size = 10, legend.text.size = 8, title.color = NULL,
  subtitle.color = "grey20", text.color = NULL,
  axis.title.color = "grey20", axis.text.color = "grey20",
  legend.title.color = "grey20", legend.text.color = "grey20")
}
\arguments{
\item{text.font}{text font}

\item{title.font}{title font}

\item{legend.font}{legend font}

\item{title.size}{title font size}

\item{text.size}{text font size}

\item{subtitle.size}{subtitle font size}

\item{axis.title.size}{axis title font size}

\item{axis.text.size}{axis text font size}

\item{legend.title.size}{legend title font size}

\item{legend.text.size}{legend text font size}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Avatar: The Last Airbender theme, Recommended font: "Slayer"
}
\details{
DETAILS
}
