% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_thelastairbender.R
\name{theme_theLastAirbender}
\alias{theme_theLastAirbender}
\title{theme_theLastAirbender}
\usage{
theme_theLastAirbender(font.text = "Slayer", font.title = "Slayer",
  font.legend = "Slayer", title.size = 18)
}
\arguments{
\item{font.text}{PARAM_DESCRIPTION, Default: 'Slayer'}

\item{font.title}{PARAM_DESCRIPTION, Default: 'Slayer'}

\item{font.legend}{PARAM_DESCRIPTION, Default: 'Slayer'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Avatar: The Last Airbender theme
}
\details{
DETAILS
}
