% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_thelastairbender.R
\name{theme_theLastAirbender}
\alias{theme_theLastAirbender}
\title{Avatar: The Last Airbender theme (deprecated)}
\usage{
theme_theLastAirbender(
  text.font = NULL,
  title.font = NULL,
  legend.font = NULL,
  title.size = 14,
  text.size = 10,
  subtitle.size = 12,
  axis.title.size = 10,
  axis.text.size = 8,
  legend.title.size = 10,
  legend.text.size = 8,
  title.color = NULL,
  subtitle.color = "grey20",
  text.color = NULL,
  axis.title.color = "grey20",
  axis.text.color = "grey20",
  legend.title.color = "grey20",
  legend.text.color = "grey20",
  legend.position = "bottom",
  ticks = FALSE
)
}
\arguments{
\item{text.font}{text font, Default: NULL}

\item{title.font}{title font, Default: NULL}

\item{legend.font}{legend font, Default: NULL}

\item{title.size}{title font size, Default: 14}

\item{text.size}{text font size, Default: 10}

\item{subtitle.size}{subtitle font size, Default: 12}

\item{axis.title.size}{axis title font size, Default: 10}

\item{axis.text.size}{axis text font size, Default: 8}

\item{legend.title.size}{legend title font size, Default: 10}

\item{legend.text.size}{legend text font size, Default: 8}

\item{title.color}{title color, Default: NULL}

\item{subtitle.color}{subtitle.color, Default: "grey20"}

\item{text.color}{text color, Default: NULL}

\item{axis.title.color}{axis title color, Default: "grey20"}

\item{axis.text.color}{axis text color, Default: "grey20"}

\item{legend.title.color}{legend title color, Default: "grey20"}

\item{legend.text.color}{legend text color, Default: "grey20"}

\item{legend.position}{legend position, Default: "bottom"}

\item{ticks}{add axis ticks, Default: FALSE}
}
\description{
Avatar: The Last Airbender theme, Recommended font: "Slayer"
}
\seealso{
[ggplot2::theme]
}
